+++ src/structure/systems.rs
@@ -33,8 +33,8 @@ pub fn spawn_base(
                     ..default()
                 },
                 transform: Transform::from_xyz(
-                    tile_pos_to_world((GRID_COLS as f32 / 2.0) as u32),
-                    tile_pos_to_world((GRID_ROWS as f32 / 2.0) as u32),
+                    tile_pos_to_world(GRID_COLS / 2.0),
+                    tile_pos_to_world(GRID_ROWS / 2.0),
                     STRUCTURE_Z_INDEX,
                 ),
                 ..default()
